+++ kaithem/src/thirdparty/upnpclient/upnp.py
@@ -179,8 +179,8 @@ class Device(CallActionMixin):
                 findtext('SCPDURL'),
                 findtext('eventSubURL')
             )
-            self._log.debug(
-                '%s: Service %r at %r', self.device_name, svc.service_type, svc.scpd_url)
+            # self._log.debug(
+            #     '%s: Service %r at %r', self.device_name, svc.service_type, svc.scpd_url)
             self.services.append(svc)
             self.service_map[svc.name] = svc
 
@@ -218,13 +218,13 @@ class Service(CallActionMixin):
         self.statevars = {}
         self._log = _getLogger('Service')
 
-        self._log.debug('%s url_base: %s', self.service_id, self._url_base)
-        self._log.debug('%s SCPDURL: %s', self.service_id, self.scpd_url)
-        self._log.debug('%s controlURL: %s', self.service_id, self._control_url)
-        self._log.debug('%s eventSubURL: %s', self.service_id, self._event_sub_url)
+        # self._log.debug('%s url_base: %s', self.service_id, self._url_base)
+        # self._log.debug('%s SCPDURL: %s', self.service_id, self.scpd_url)
+        # self._log.debug('%s controlURL: %s', self.service_id, self._control_url)
+        # self._log.debug('%s eventSubURL: %s', self.service_id, self._event_sub_url)
 
         url = urljoin(self._url_base, self.scpd_url)
-        self._log.debug('Reading %s', url)
+        # self._log.debug('Reading %s', url)
         resp = requests.get(
           url,
           timeout=HTTP_TIMEOUT,